move the constant file for LoginPortal and SignUpPortal
